feat : 좌표 소수점 지원하도록 수정 - short -> float 변경

+++ src/main/java/com/hoo/aoo/admin/application/service/Metadata.java
@@ -22,9 +22,9 @@ public record Metadata(
     public record Room(
             String formName,
             String name,
-            Integer x,
-            Integer y,
-            Integer z,
+            Float x,
+            Float y,
+            Float z,
             Integer width,
             Integer height
     ) {
